Parallel processing code.
This file contains the function to launch the kernel on the device as well as the kernel itself. The kernel converts the image to greyscale using a weighted conversion. The function creates the blockSize and gridSize before launching the kernel. The blockSize is currently set to include a whole row meaning that it launches numCols threads per block. 

Runtime = 0.028 ms